Twitter::Cursor is Enumerable; just call to_a
There's certainly no need for you to write your own recursive function.
The whole point of the Twitter::Cursor class is to encapsulate the logic
of how to iterate through results.

That's why Twitter::Cursor#next_cursor is a private method. As a rule,
you should never need to use the #attrs method on any Twitter object.
I've made the raw data available for contingency purposes but it's
almost always an indication that you're doing something wrong.